+++ pyaerocom/io/read_aasetal.py
@@ -9,7 +9,7 @@ from pyaerocom.helpers import get_tot_number_of_seconds
 from pyaerocom.io.readungriddedbase import ReadUngriddedBase
 from pyaerocom.stationdata import StationData
 from pyaerocom.ungriddeddata import UngriddedData
-from pyaerocom.units_helpers import convert_unit
+from pyaerocom.units.units_helpers import convert_unit
 
 logger = logging.getLogger(__name__)
