+++ pyaerocom/io/read_ebas.py
@@ -48,7 +48,7 @@ from pyaerocom.molmasses import get_molmass
 from pyaerocom.stationdata import StationData
 from pyaerocom.tstype import TsType
 from pyaerocom.ungriddeddata import UngriddedData
-from pyaerocom.units_helpers import get_unit_conversion_fac
+from pyaerocom.units.units_helpers import get_unit_conversion_fac
 
 logger = logging.getLogger(__name__)
